option to request acceptance of terms and conditions
if variable SELLYOURSAAS_TERMSANDCONDITIONS is set with name of T&C page (e.g. terms.php),
a checkbox is shown to request acceptance of the terms and conditions. Accepting this checkbox
is then mandatory.

The message for the checkbox is stored under the key WhenRegisteringYouAccept (same as for the
dolicloud specific mechanism to show T&C pages).